pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "trolo"
version = "0.1.0.dev1"
description = "CLI tool for D-FINE model training and inference"
requires-python = ">=3.8"
dependencies = [
    "click>=8.0.0",
    "torch>=2.0.0",
    "torchvision",
    "onnxruntime",
    "PyYAML",
    "opencv-python",
    "Pillow",
]

[project.scripts]
trolo = "trolo.cli.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["trolo*"]

[tool.setuptools.package-data]
trolo = ["configs/**/*.yml", "configs/**/*.yaml"]
